\name{GHmixedeffect-package}
\alias{GHmixedeffect-package}
\alias{GHmixedeffect}
\docType{package}
\title{
\packageTitle{GHmixedeffect}
}
\description{
\packageDescription{GHmixedeffect}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GHmixedeffect}
\packageIndices{GHmixedeffect}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GHmixedeffect}

Maintainer: \packageMaintainer{GHmixedeffect}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
